Port website over to a Laravel 5.4 project.
This commit has a lot of changes. I ported the old website over to Laravel, and then redeisgned much of the website. Mostly the content is still the same, however it has a totally different look and feel that I feel is more modern and just mroe pleasing on the eye.

There are big plans for this in the future as well in refining everythign and making the site more dynamic with an admin protal and the site content to be mostly dynamic and editable from teh admin portal.