Contact email: Fix the API call to send the contact email.
Cleaned up the logic for formatting the JSON API response.
Switched to using a bootstrap modal instead of snackbarjs.
Make sure the button is disabled until the API response comes back.
Handle any error response from the API and display the content in the modal.
Configured the Laravel exception handler to always return JSON on API requests in the event of an exception.
Throttle the contact email API requests to 1 per minute.