JSON Sanity: Bring some sanity to JSON responses
By default, exceptions are rendered as HTML, which is undesireable for requests expecting JSON.
I had implemented a check in the exception handler to check that the request expected JSON and reutrn a JSON response.
However there was some issues with exceptions being thrown that did not have the getStatusCode function and ones that can return an empty message, so I added some checks to make sure we should always get back some sort of JSON response.

Also any TokenMismatchException thrown was rendered as HTML, and it's exception did not have a message or status, so it would just return a empty erro message. I modified the CSRF middleware to check if the request expects JSON, and if so return a 403 JSON response to inform the end-user of a CSRF token mismatch.