recaptcha: Add Google's reCAPTCHA to the contact form
Also use a custom FormRequest for the contact email route with validation.
Extend the throttle to 5 request per minute on the contact email route.
Simplfy contact email controller.
Refactor AJAX request and response handling in the JavaScript. We should be able to handle any response given by the server and display the response in a freidnly format in the modal now. Also the JavaScript won't even try to make the cotnact request if no Google reCAPTCHA JavaScript is present.
Also validation errors from the server will be displayed in the modal.