Feature: Refactor LastModified service.
I took a naive approach to get the last modifed file time for my application directory. This recursed the whole project structure (including the git directory, node_modules, vendor, etc) and then I filtered out some of the directories I didn't want to take into account on the last modified time.

Also having the service bound as a function made it execute twice, once during bootstrapping and another whenit was actually used.

This function had around a half a second run time on my local machine which is way to much.

So my new approach was to first specify a list of directories to traverse. This includes the app folder, bootstrap, config, database, public, resources and the tests folders. I had to have a separate case to handle top level files in the base path, since using the RecursiveDirectoryIterator on the base path would lead to recursing down all folders like before. So for each directory specified I recurse into it and compare all file MTimes using the SplFileInfo objects. I then use a DirectoryIterator on the base path and get the top level files and compare them. Once I have the greatest file modified time, I cache that result for 30 minutes and reutrn the Carbon instance.

Without factoring in the caching, this algorithm went from 1/2 second to around 5ms. With caching in place (file caching driver on local machine) we are in the 0.3ms range. Much much better.

Plus the service is now in a class, which will make testing it easier. I was going to use the Laravel Filesystem object, but I had to many issues controlling the directory recursion, and I can test this using vfsStream to mock out the filesystem hopefully.