feature: Add route for Expect-CT header failures
I will be adding the Expect-CT header to the nginx config and it is
recomended to enable it in report only mode and log the failures
before enforcing the Expect-CT header. This route can be used to
send Expect-CT failure reports to and will send out an email
detailing the report as well as log the report.