feature/unit-tests: Add unit tests for LastModified service
Also refactor the functionality of the LastModified service.
It is now a deferred service.
It also has an interface, so where it is used can rely on the interface
and not a concrete class.

Also I added a view composer for the partials.lastmodified template to
expose the value from the LastModidifed@getLastModifiedTime function to
that partial view. Now anywhere it is included, it will be provided that
value.

Also added the cache prefix to the environment file.

Added the ability to configure the cache key for the last modified
timestamp in the config (and in the env file) as well as the directories
it will traverse.

Also added some feature tests to assert the index page returns a 200 and
an invalid route returns a 404.